collapsing dataframes (#13)
* prep for develop_mode

* change to development mode

* switch magic 13 to proper column name

* adding function "dedupeMuts" for collapsing rows in sigmut dataframe

* split chunk "processa_and_match_snvs_to_signature_mutations" in chunk for processing the sigmut dataframe and filtering the variant dataframe

* adding procedure for deduplicating the sigmut dataframe and grouping the variant names per mutation

* change "left_join" so the variant dataframe is now joined with the deduplicated version of the sigmut dataframe

* fixed typo in line 312

* change "getIdentCols" function to the refined function "dedupe"

* adapting chunk "collapse_identicals" to use "dedupe" function

* adapting chunk "getting_unique_muts_bulk", match.df is already unique now and doesn't need that extra "distinct()" step anymore

* fix typo in line 375

* comment development chunks again

* Dev rendersite (#11)

* added rendercite rules to snakemake

adapted snakefile and rendercite to communicate with report_scripts

* dummy Kraken and Krona output files - to test rendersite

updated timecourse page to be in nav_bar

* moved two report scripts of vic into report_scripts/

added qc-report and its params  to rendersite generation

* added rule to snakefile and from .rmd to .Rmd

* changed test file to new sample csv file

* modified _settings and suppressed warnings

added snakemake rules for rendersite and commented out the ones which are not needed any more

* renamed variantreport

inserted variantreport to communicate with snakefile and generaterenderFiles. Also added the rule to the snakefile

* changed test paths in files and added plot snippets from rafael

deletet vep file because old

changed settings file so timecourse can read in the params and added test_vep file

added bubble figure try

 deleted commented lines for pull request

* adapted st reports are generated via render_site

removed commented lines for pull request

* added krona to kraken report

* deleted krona report, this will be handled by kraken2_report

changed name of kraken2-report to taxonomic classification

added index.rmd

* changed paste0 to file.path()

Co-authored-by: Dohmen <jan.dohmen@googlemail.com>

* changing structure, adding metadata

* adapting to renderStie structure

* it's still a mixed structure with some development-testing parts left

adapted snakefile and rendercite to communicate with report_scripts

* dummy Kraken and Krona output files - to test rendersite

updated timecourse page to be in nav_bar

* moved two report scripts of vic into report_scripts/

added qc-report and its params  to rendersite generation

* added rule to snakefile and from .rmd to .Rmd

* changed test file to new sample csv file

* modified _settings and suppressed warnings

added snakemake rules for rendersite and commented out the ones which are not needed any more

* renamed variantreport

inserted variantreport to communicate with snakefile and generaterenderFiles. Also added the rule to the snakefile

* changed test paths in files and added plot snippets from rafael

deletet vep file because old

changed settings file so timecourse can read in the params and added test_vep file

added bubble figure try

 deleted commented lines for pull request

* adapted st reports are generated via render_site

removed commented lines for pull request

* added krona to kraken report

* deleted krona report, this will be handled by kraken2_report

changed name of kraken2-report to taxonomic classification

added index.rmd

* changed paste0 to file.path()

Co-authored-by: Dohmen <jan.dohmen@googlemail.com>

* uncomment test env

* comment renderSite env

* uncomment test env

* adding "dedupeMuts" again

I lost the function in the recent merging process

* adding start of writing var_output

* adding sample name var

* adding comment for input args

* adding skeleton for output.csv

dataframe skeleton and first functions are added for creating the variant_output.

* adding variable for csv output directory

* writing  data_variant_plot.csv

added part for writing the csv file which is used to plot the variant abundance.

* switching off/commenting dev chunks

* remove redundant "samplename"

samplename variable changed to sample_name since it was already defined earlier

* adding csv-writing part again

Idk why but for some reason it got lost during the merging process

* prep for develop_mode

* change to development mode

* switch magic 13 to proper column name

* adding function "dedupeMuts" for collapsing rows in sigmut dataframe

* split chunk "processa_and_match_snvs_to_signature_mutations" in chunk for processing the sigmut dataframe and filtering the variant dataframe

* adding procedure for deduplicating the sigmut dataframe and grouping the variant names per mutation

* change "left_join" so the variant dataframe is now joined with the deduplicated version of the sigmut dataframe

* fixed typo in line 312

* change "getIdentCols" function to the refined function "dedupe"

* adapting chunk "collapse_identicals" to use "dedupe" function

* adapting chunk "getting_unique_muts_bulk", match.df is already unique now and doesn't need that extra "distinct()" step anymore

* fix typo in line 375

* adapting to renderStie structure

* it's still a mixed structure with some development-testing parts left

adapted snakefile and rendercite to communicate with report_scripts

* dummy Kraken and Krona output files - to test rendersite

updated timecourse page to be in nav_bar

* moved two report scripts of vic into report_scripts/

added qc-report and its params  to rendersite generation

* added rule to snakefile and from .rmd to .Rmd

* changed test file to new sample csv file

* modified _settings and suppressed warnings

added snakemake rules for rendersite and commented out the ones which are not needed any more

* renamed variantreport

inserted variantreport to communicate with snakefile and generaterenderFiles. Also added the rule to the snakefile

* changed test paths in files and added plot snippets from rafael

deletet vep file because old

changed settings file so timecourse can read in the params and added test_vep file

added bubble figure try

 deleted commented lines for pull request

* adapted st reports are generated via render_site

removed commented lines for pull request

* added krona to kraken report

* deleted krona report, this will be handled by kraken2_report

changed name of kraken2-report to taxonomic classification

added index.rmd

* changed paste0 to file.path()

Co-authored-by: Dohmen <jan.dohmen@googlemail.com>

* comment renderSite env

* adding "dedupeMuts" again

I lost the function in the recent merging process

* adding start of writing var_output

* adding sample name var

* adding comment for input args

* adding skeleton for output.csv

dataframe skeleton and first functions are added for creating the variant_output.

* adding variable for csv output directory

* writing  data_variant_plot.csv

added part for writing the csv file which is used to plot the variant abundance.

* switching off/commenting dev chunks

* remove redundant "samplename"

samplename variable changed to sample_name since it was already defined earlier

* adding csv-writing part again

Idk why but for some reason it got lost during the merging process

Co-authored-by: mirifax <45397751+mirifax@users.noreply.github.com>
Co-authored-by: Dohmen <jan.dohmen@googlemail.com>