Avoid runtime error during merge, no headers in table
For unique column names during merge, the ".x" or ".y" suffix is added
but this is not sufficient for repeated merges when one side alread has
a column name that ends with ".x" or ".y". This patch adds unique
column names to avoid such clashes.

Since headers are not informative, neither with this patch nor before
when these were just variants of "V2", the non-sensical headers should
not be part of the table as it is written into a file. Also, the tools
downstream cannot read the table with the default export format that
lists headers but drops the first field where there are row.names.